+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - The In/Out of Garrison Sensemaker now creates a finding type upon processing data
+- The Incursion Sensemaker now creates a finding type upon processing data
 
 ### Changed
 - Grimlock now uses SDK version 12.0.0